// src/__generated__/scheduleRunsRelay_query.graphql.ts
export type ScriptExecutionStatus = "FAILED" | "QUEUED" | "RUNNING" | "SUCCESS" | "%future added value";
import type { Instant } from "../lib/graphql-scalars";
import type { FragmentRefs } from "relay-runtime";
export type scheduleRunsRelay_query$data = {
  readonly scheduleRuns: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly dispatchedAt: Instant;
        readonly executionId: string;
        readonly id: string;
        readonly initiator: {
          readonly email: string | null | undefined;
          readonly firstName: string | null | undefined;
          readonly id: string;
          readonly image: {
            readonly hash: string | null | undefined;
            readonly imageUrl: string | null | undefined;
          } | null | undefined;
          readonly lastName: string | null | undefined;
          readonly status: string | null | undefined;
        } | null | undefined;
        readonly respondedMachineCount: number;
        readonly status: ScriptExecutionStatus;
        readonly totalMachineCount: number;
      };
    }>;
    readonly filteredCount: number;
    readonly pageInfo: {
      readonly endCursor: string | null | undefined;
      readonly hasNextPage: boolean;
    };
  };
  readonly " $fragmentType": "scheduleRunsRelay_query";
};
export type scheduleRunsRelay_query$key = {
  readonly " $data"?: scheduleRunsRelay_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"scheduleRunsRelay_query">;
};